Refactor to packages.
