+++ src/main/java/ooaddemo/controller/ValidatingController.java
@@ -1,4 +1,7 @@
-package ooaddemo;
+package ooaddemo.controller;
+
+import ooaddemo.domain.SeverityLevel;
+import ooaddemo.message.DecoratingMessage;
 
 public abstract class ValidatingController {
     public void log(DecoratingMessage message, SeverityLevel severity) {
