+++ src/main/java/ooaddemo/domain/SeverityLevel.java
@@ -1,4 +1,4 @@
-package ooaddemo;
+package ooaddemo.domain;
 
 public enum SeverityLevel {
     ERROR, WARNING, DEBUG
