+++ src/main/java/ooaddemo/filter/MessageFilter.java
@@ -1,4 +1,7 @@
-package ooaddemo;
+package ooaddemo.filter;
+
+import ooaddemo.message.DecoratingMessage;
+import ooaddemo.domain.SeverityLevel;
 
 public interface MessageFilter {
     boolean filter(DecoratingMessage message, SeverityLevel severity);
