+++ src/main/java/ooaddemo/message/DecoratingMessage.java
@@ -1,4 +1,4 @@
-package ooaddemo;
+package ooaddemo.message;
 
 public interface DecoratingMessage {
     String getDecoratedMessage();
