+++ src/main/java/ooaddemo/printer/Printer.java
@@ -1,4 +1,4 @@
-package ooaddemo;
+package ooaddemo.printer;
 
 public interface Printer {
     void print(String message);
